Fix a flaky test case in find.test.
The -newer check was failing about once in three times on my machine.
'stat' showed that 'file' and 'link' can have the exact same timestamp,
probably due to some metadata buffering in the kernel.

Also remove the testdata cleanup, since it makes it harder to debug, and
that's the job of the harness.